Add decorators for api
Add two common decorators, update_token and notify subject.

Update_token is to update token from the return data from api caller.
This is a SOP, so we move it to the module level to reduce the code size
and keep it simple.

Notify subject is a utilization decorator which helps to notify speical
subject which is already subscribed.