Refactor the execute method
We want to use `getattr` to auto get the method by combine the function
name. It can reduce the code we have to write :/

Execute method will find the method by named `handle_{command}`. For
example, if we have command `battle`, then the method named
`handle_battle` will be found and executed.